[build-script] Don't use host to decide SDK target
Prior to this commit, SDK targets were determined based on the host
system name and processor. For example, `swift-stdlib-linux-armv7`
would only be added as a CMake build target if the host system name was
"Linux", and if the host system processor was "armv7l".

Instead, this commit configures the CMake executable format and other
parameters based on the host system, and allows any SDK target
specified in the `build-script-impl` file's `STDLIB_DEPLOYMENT_TARGETS`
list to be built on any platform.

Of course, not every target can be built on every platform. SDKs
targeting 'watchos-armv7k' will undoubtedly fail to compile on Linux
host machines. Still, the build script no longer makes assumptions on
which host system can cross-compile which targets.